fix linting issues and merge utils files
