+++ test/cli/test_models.py
@@ -16,27 +16,33 @@ class TestTaskObject:
 
 
 class TestDaylistObject:
+    @pytest.fixture()
     @staticmethod
-    def setup_tasks(titles=["task 1", "task 2"], durs=[60, 120]) -> list[TaskCLI]:
-        if len(titles) != len(durs):
-            raise ValueError("Number of provided titles and estimates don't match")
-        tasks = []
-        for title, dur in zip(titles, durs):
-            tasks.append(TaskCLI(title=title, estimate=dur))
-
+    def initial_tasks() -> list[TaskCLI]:
+        tasks = [
+            TaskCLI(title="one", estimate=timedelta(minutes=20)),
+            TaskCLI(title="two", estimate=timedelta(minutes=20)),
+            TaskCLI(title="three", estimate=timedelta(minutes=20)),
+        ]
         return tasks
 
     def test_total_estimate(self):
         task_times = [10, 120]
-        tasks = self.setup_tasks(durs=task_times)
+        tasks = [
+            TaskCLI(title="one", estimate=timedelta(seconds=task_times[0])),
+            TaskCLI(title="two", estimate=timedelta(seconds=task_times[1])),
+        ]
         test_list = DaylistCLI(pending_tasks=tasks)
 
         # Assert tasks times are summed into the estimate
         assert test_list.total_estimate().total_seconds() == sum(task_times)
 
     def test_total_estimate_with_done(self):
-        task_times = [10, 120]
-        tasks = self.setup_tasks(durs=task_times)
+        task_times = [60, 60]
+        tasks = [
+            TaskCLI(title="one", estimate=timedelta(seconds=task_times[0])),
+            TaskCLI(title="two", estimate=timedelta(seconds=task_times[1])),
+        ]
         test_list = DaylistCLI(pending_tasks=tasks)
 
         # Expect done tasks do not contribute to total estimate
@@ -56,7 +62,7 @@ class TestDaylistObject:
         test_list = DaylistCLI(expiry=sample_datetime)
         assert test_list.is_expired() == expected
 
-    def test_add_task(self):
+    def test_add_task_to_empty(self):
         test_list = DaylistCLI()
         assert len(test_list.pending_tasks) == 0
 
@@ -66,34 +72,41 @@ class TestDaylistObject:
         assert len(test_list.pending_tasks) == 1
         assert test_list.pending_tasks[0].title == test_title
 
-    def test_remove_task(self):
-        tasks = self.setup_tasks()
-        test_list = DaylistCLI(pending_tasks=tasks)
+    def test_add_task_to_others(self, initial_tasks):
+        test_list = DaylistCLI(pending_tasks=initial_tasks)
+        assert len(test_list.pending_tasks) == len(initial_tasks)
 
-        assert len(test_list.pending_tasks) == len(tasks)
+        test_title = "hello"
+        test_delta = timedelta(minutes=30)
+        test_list.add_task(title=test_title, estimate=test_delta)
+        assert len(test_list.pending_tasks) == len(initial_tasks) + 1
+
+    def test_remove_task(self, initial_tasks):
+        test_list = DaylistCLI(pending_tasks=initial_tasks)
+
+        assert len(test_list.pending_tasks) == len(initial_tasks)
         test_list.remove_task(1)
-        assert len(test_list.pending_tasks) == (len(tasks) - 1)
+        assert len(test_list.pending_tasks) == (len(initial_tasks) - 1)
         test_list.remove_task(0)
-        assert len(test_list.pending_tasks) == (len(tasks) - 2)
+        assert len(test_list.pending_tasks) == (len(initial_tasks) - 2)
 
-    @pytest.mark.parametrize("bad_index", [10, -1])
-    def test_remove_task_bad_index(self, bad_index):
-        test_list = DaylistCLI(pending_tasks=self.setup_tasks())
+    @pytest.mark.parametrize("bad_index", [100, -1])
+    def test_remove_task_bad_index(self, bad_index, initial_tasks):
+        test_list = DaylistCLI(pending_tasks=initial_tasks)
         with pytest.raises(IndexError):
             test_list.remove_task(bad_index)
 
-    def test_complete_task(self):
-        tasks = self.setup_tasks()
-        test_list = DaylistCLI(pending_tasks=tasks)
+    def test_complete_task(self, initial_tasks):
+        test_list = DaylistCLI(pending_tasks=initial_tasks)
 
-        assert len(test_list.pending_tasks) == len(tasks)
+        assert len(test_list.pending_tasks) == len(initial_tasks)
         test_list.complete_task(1)
-        assert len(test_list.pending_tasks) == (len(tasks) - 1)
+        assert len(test_list.pending_tasks) == (len(initial_tasks) - 1)
         test_list.complete_task(0)
-        assert len(test_list.pending_tasks) == (len(tasks) - 2)
+        assert len(test_list.pending_tasks) == (len(initial_tasks) - 2)
 
-    @pytest.mark.parametrize("bad_index", [10, -1])
-    def test_complete_task_bad_index(self, bad_index):
-        test_list = DaylistCLI(pending_tasks=self.setup_tasks())
+    @pytest.mark.parametrize("bad_index", [100, -1])
+    def test_complete_task_bad_index(self, bad_index, initial_tasks):
+        test_list = DaylistCLI(pending_tasks=initial_tasks)
         with pytest.raises(IndexError):
             test_list.complete_task(bad_index)
